Integer division makes apix 0

+++ project.py
@@ -69,7 +69,7 @@ def main(args):
         log.error("Some shifts are too large to crop (maximum crop is %d)" % (args.size - 2 * maxshift))
         return 1
 
-    apix = star.calculate_apix(df) * args.size / (f3d.shape[0] // args.pfac - 1)
+    apix = star.calculate_apix(df) * np.double(args.size) / (f3d.shape[0] // args.pfac - 1)
 
     sz = f3d.shape[0] // args.pfac - 1
     sx, sy = np.meshgrid(np.fft.rfftfreq(sz), np.fft.fftfreq(sz))
@@ -126,7 +126,7 @@ def project(f3d, p, s, sx, sy, a, pfac=2, apply_ctf=False, size=None, flip_phase
     f2d = vop.interpolate_slice_numba(f3d, orient, pfac=pfac, size=size)
     f2d *= pshift
     if apply_ctf or flip_phase:
-        apix = star.calculate_apix(p) * size / (f3d.shape[0] // pfac - 1)
+        apix = star.calculate_apix(p) * np.double(size) / (f3d.shape[0] // pfac - 1)
         c = ctf.eval_ctf(s / apix, a,
                          p[star.Relion.DEFOCUSU], p[star.Relion.DEFOCUSV],
                          p[star.Relion.DEFOCUSANGLE],
